refactor(setup): extract restore wizard routes
